use uint32_t for all config values

+++ src/EffectDevices/WavePWMDevice.h
@@ -34,7 +34,7 @@ class WavePWMDevice : public EffectDevice, public HighPowerOffAware {
     this->wavePWM = new WavePWM();
     this->pin = pin;
     // A hard limit, for example to run a 12V motor at 48V, maxPWM has to be 63.
-    this->pwmLimitFactor = maxPWM / 255;
+    this->pwmLimitFactor = static_cast<float>(maxPWM) / 255.0f;
     pinMode(pin, OUTPUT);
   }
